+++ src/main/php/valuereader/CommonValueReader.php
@@ -13,6 +13,7 @@ use stubbles\input\filter\JsonFilter;
 use stubbles\input\filter\PasswordChecker;
 use stubbles\input\filter\range\DateRange;
 use stubbles\input\filter\range\DatespanRange;
+use stubbles\input\filter\range\SecretMinLength;
 use stubbles\input\filter\range\StringLength;
 use stubbles\input\filter\range\NumberRange;
 use stubbles\values\Secret;
@@ -71,10 +72,10 @@ interface CommonValueReader
     /**
      * read as secret
      *
-     * @param   \stubbles\input\filter\range\StringLength  $length
+     * @param   \stubbles\input\filter\range\SecretMinLength  $length
      * @return  \stubbles\values\Secret
      */
-    public function asSecret(StringLength $length = null): ?Secret;
+    public function asSecret(SecretMinLength $length = null): ?Secret;
 
     /**
      * read as text value